hackernews/requirements.txt
amqp==5.1.1
APScheduler==3.10.0
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.5
ez-setup==0.9
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
prompt-toolkit==3.0.36
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.1
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.3.0
wrapt==1.14.1
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
